build: update dependency to latest v13 release and fix errors (#1516)
* build: update dependency to latest v13 release and fix errors

The latest v13 release of the language service removes the option to
disable Ivy from the `PluginConfig`. This commit also removes it from
the session configuration, but does not add a workaround to continue
supporting View Engine. That will need to be done before the v13
release.

In addition, the `esModuleInterop` option was enabled, which matches the
angular/angular repo. This addresses another build issue due to default
import/exports.

* fixup! build: update dependency to latest v13 release and fix errors