refactor: support resolving ngcc from APF v13 output
Angular v13 version comes with the APF v13
output where all FW packages are strict ESM. This required some
changes in the ngcc lookup logic to properly work with strict ESM
and the new `package.json` `exports` field.

This is needed in v12 because a workspace may include projects that are
using different versions of Angular. Because we resolve ngcc from the
node modules of the application code, we need to support both formats.