fix(server): return the right range for the original source file of DTS (#1604)
When the user tries to go to the definition that is in the DTS file,
the Angular LS will map it to the original source file. Because of
not updating the `scriptInfo`, the Angular LS returns the wrong range
of the definition.

(cherry picked from commit 9a5ecd92cf617259e474fc23461cb586cc6ead76)