refactor: use built-in `WorkDoneProgressReport` to report NGCC progress (#1614)
Now the Angular extension implements the `WorkDoneProgressReport`
by itself on the client-side to support the multi-report progress
with different `tsconfig.json`, to show an error message, to
dispose of the report progress, but it's already a built-in feature
that can be used on the service side.

On the service side, the built-in `WorkDoneProgressReport` will
attach a unique token to the different report progress and clean
the progress when the client is closing.

On the service side, If an error report when running NGCC, the LSP
also supports showing an error message in VSCode, but the ability to
open the output channel is not supported(now the LSP only supports
to open the document URI). This PR will only need to provide it.

So the custom `WorkDoneProgressReport` can be removed on the client side.

(cherry picked from commit 51578720168bc347ae00dd8bc0aafd3dacca9034)