feat: Add option to disable ngcc (#1620)
Running ngcc is required for gathering metadata information of libraries
not published with Ivy instructions. However, ngcc can also be run
manually _outside_ of the extension. Developers may often be running a dev
server which does this step so doing it in the extension as well is
redundant. In addition, there have been issues with running it in VSCode
(#1353, #1444, and potentially other Nx related issues).

This option gives developers the ability to manually opt out of running
ngcc from the extension if it is proving to be problematic for their
setup.

Fixes #1353