feat: add command to run ngcc manually
Adds a command to manually run ngcc for a given file. This commit also
refactors the `disableNgcc` option to indicate that it's disabling the
_automatic_ `ngcc` run but it can still be triggered manually with the
new command.

The `disableAutomaticNgcc` option and new command to manually run ngcc
are complementary. We have found that the extension can be overzealous
in automatically running ngcc, especially with solution-style project
configurations (common practice in Nx projects). If a dev server is not
active to rerun ngcc when needed, then the new command gives an easy way
to do that when the `disableAutomaticNgcc` option is set.

Fixes #991