Fix(server): Ensure open templates keep their corresponding projects open
VSCode/TSServer implements an optimization to close projects which no
longer have any open files in the editor (see `cleanUpAfterOpeningFile`
in TypeScript's `ProjectService`). This optimization currently results
in closing projects when only a template file is open because the
template files are not sufficiently tracked (since they're non-TS
files).

A previous attempt to fix this issue was implemented in #1461 but does
not work for solution-style `tsconfig` projects because when creating
the external project, it only counts towards the
`externalProjectRefCount` if the name of the tsconfig is `tsconfig.json`
or `jsconfig.json` (See call to `getBaseConfigFileName` in
`openExternalProjects`).

Instead, this fix adds the template files to the project's `rootFiles`.
This is sufficient to ensure the `hasOpenRef` check passes and the
template's project is retained when navigating to a file that exists in
a different project.

Fixes #1623
Fixes #876