fix(server): Ensure all projects that finish ngcc re-enable language service features
The implementation of the ngcc queue for re-enabling project language
service's is incorrect. It unconditionally removes projects from the
queue when the ngcc process finishes even if it finished out of order
and is waiting for the first project in the queue to finish. This
results in later projects never getting the language service re-enabled
when the earlier project(s) finally finish ngcc.

This fix works towards stabilizing the extension for solution-style projects.

Related to #876