feat(extension): Add option to force strict templates
This feature request is to provide an extension option to override the project's
`angularCompilerOptions`. This would allow the language service to enable `strictTemplates`.

With the release of v12, the Ivy-based Language Service became the default.
With that, the information provided by the extension follows what the application
has configured the compiler to interpret (see Alex's
[blog post](https://blog.angular.io/under-the-hood-of-the-language-service-ab763c26f522)
for more information on why this was done). As a result, developers upgrading from
previous versions that have apps which do not have strict templates enabled often
experience a loss of information from the extension. It also may not be feasible
for large projects to enable `strictTemplates` immediately due to an abundance of new errors that would need to be fixed.

Importantly, this would likely result in the language service producing diagnostics
that are not produced when running/compiling the application. However, this may
still be a more ideal state than the current experience (which would be simply
not getting _any_ useful information for lots of locations in the template which
require stricter type checking options).

resolves #1418