fix(server): send diagnostic range to the Angular language service when fixing code errors (#1747)
This will make the fix error code simple, the range the fix error code get is always
the start and end of the diagnostic. For example, `<span>{{title}}</span>`, the `title` is
the missing member of component. When the range the user selects is from the start of `<span>`
to the end of `</span>`, the fix error code needs to check the start of the diagnostic.