fix(server): resolve tsdk correctly when settings specify a relative location
Previously, we only supported absolute paths for `tsdk` specified in the
settings.json. However, with Yarn PnP, this might instead be a path
relative to the project/workspace. For example, in a a project using
Yarn PnP, the tsdk setting might be

```
"typescript.tsdk": ".yarn/sdks/typescript/lib",
```

The user would then run `yarn dlx @yarnpkg/sdks vscode` and the
typescript modules would then exist in the project folder
`.yarn/sdks/typescript/...`. We should then resolve tsdk as a relative
path from the given probe locations, just like we do when looking for
other packages in a regular project (in the `node_modules`, relative to
the given probe locations).

Fixes #1748