feat(server): provide folding ranges for inline templates (#1779)
* test(server): Remove spec for view engine, which is no longer supported

To support legacy view engine, we install an old version of the language
server into the vsix.

* build: remove unused esbuild.js

The esbuild configuration is now managed in the BUILD.bazel rules.

* feat(server): provide folding ranges for inline templates

Following the embedded language support documentation, this commit provides
folding ranges for inline templates in typescript files.
This feature is provided in the language server so other editors that
use the @angular/language-server package can make use of the feature as
well.

https://code.visualstudio.com/api/language-extensions/embedded-languages

fixes #852

* fixup! feat(server): provide folding ranges for inline templates

* fixup! feat(server): provide folding ranges for inline templates

* fixup! feat(server): provide folding ranges for inline templates