build: do not manage package version in Starlark file
The package/project version should be managed via the `package.json`
file so that existing tooling like `yarn ng-dev` does not break
with the recent Bazel migration. Tools like `ng-dev` expect a project
version to be present in the top-level `package.json`. The current
placeholder breaks the release and branching expectations.