refactor: Remove ngcc from extension
NGCC is being removed in Angular v16. The language service should not
longer attempt to run it. https://github.com/angular/angular-cli/pull/24720
As a result, we can now set the support for running in untrusted
workspaces to "true" since we don't have to execute ngcc from the
`node_modules` folder.