fix(server): Fall back to bundled TS version if specified TSDK is too old
We attempt to use the version of typescript defined in the extension
options. However, this might be too old for the required version in the
bundled Angular compiler. If this happens, we now attempt to resolve the
version from the bundled one instead (and print a warning to the output).

Fixes #1855