fix(server): Generate help message before using attempting to resolve any require statements
The `requireOverride` resolves TSServer from the command line arguments.
The `createLogger` import will end up executing this code. As a result,
we need to parse and show the help message before the logger import or
the --help option will fail due to a failure to resolve the TS
dependency.

fixes #1899