feat(syntaxes): Support inline styles as string (#1921)
* feat(syntaxes): Support inline styles as string

https://github.com/angular/angular/commit/59387ee476dff1a893a01fe5cbee3c95b93c0cdb

The above commit added support for inline styles to be a single string literal in addition
to the previous array of strings. This commit updates the syntaxes regex of the extension
to support this feature.

* feat(language-service): Enable go to definition for styleUrl

The client code has an optimization which skips making requests to the server for Angular results
when the position is outside a context that Angular would be interested in. This commit
adds the 'styleUrl' property to those properties of interest.