refactor(client): Show warning message if a project uses a newer version than extension
When the installed extension version is less than the major version of the
project, there is likely to be errors. Because new features are added in
major/minor releases, a the bundled compiler version in an old extension version
may completely fail to even parse a template if there are new syntax features.
This problem is very apparent with the new control flow syntax. If a project uses
v17 and uses the new control flow, a v16 language service will show all kinds of parse errors.

fixes #1941