fix(syntaxes): Adjust block syntax highlighting to require ( or { on same line
While our parser allows for a block name to be defined on a separate
line from the expression or brace, the syntax highlighting will not.
Because the extension may be used for older versions of Angular and we
cannot turn highlighting on and off dynamically like we can the parser,
we make the highlighting more restrictive. This should cover a majority
of use-cases where `@` is used in older versions of Angular like in
emails and _not_ begin block highlighting.

Partially addresses #1958. Will need another change to detect Angular
version and disable block parsing.