fix: Disable block syntax parsing when no project in workspace supports it
This commit adds a check to ensure a root in the workspace supports
block syntax. If not, the block syntax parsing in the compiler is
disabled. Note that the language server is shared for all roots and all projects
in the workspace. If on supports block syntax, we have to enable it.

In the future, it would be better to find a way to make this a
per-project configuration based on the version of Angular used in that
project.

fixes #1958