refactor(server): Update tsserver version log to reflect actual version used
The version that's currently logged is whatever the resolve logic gives. However,
this resolve logic actually only affects the version when the banner is used
to override the require statement. This won't happen when debugging the extension
and stumped me for quite some time since the log version was using the tsdk
I specified but I was not seeing the behavior I expected with that version
(I was instead seeing the behavior expected from the bundled version).