fix(server): Do not disable language service for solution style root tsconfig
Do not disable language service for project if it has project references. Even though
we couldn't find angular/core for this parent tsconfig, it might exist in the
referenced projects. Disabling the language service for this root project will clean up
the program for the project.
https://github.com/microsoft/TypeScript/commit/3c4c060dff2b0292d3a4d88ba0627b7657d67e7f#diff-fad6af6557c1406192e30af30e0113a5eb327d60f9e2588bdce6667d619ebd04R973-R983
We need to keep this language service and program enabled
because referenced projects don't load until we open a client
file for that project.

fixes #1959