fix(extension): Allow requests to reach the language server when in component decorator (#2067)
This commit updates the early-return logic on the client-side to allow
requests to go to the server whenever the cursor is inside the component
decorator. Prior to this change, some requests would attempt to return
early if the position was not inside the inline template region.
However, we were never able to get the tokenizing working correctly and
this has led to issues whenever there are multiple template strings
(JS template strings with \`).

This approach should give us the needed benefits of early-returns in
non-Angular contexts (`@Component` decorator should be pretty rare) while
being more permissive and correct.

fixes #2064