fix: replace the closing symbol for inputs/outputs. (#2142)
The Angular Language Service does not return `InsertReplaceEdit`.
There is no need to allow the developer to choose how to insert the completion.

For example, `<button (c|) />`.
                      ^^__________Insert edit
                      ^^ ^________Replace edit

If the LS returns the `InsertReplaceEdit` as shown above, selecting "Insert" by the developer
results in `(click)="")`, and selecting "Replace" results in `(click)=""`.

Now in the vscode, the default `editor.suggest.insertMode` value for HTML is `Replace`, for
ts is `Insert`, So this leads to a bug in the ts file.

Fixes https://github.com/angular/vscode-ng-language-service/issues/2137