fix(language-server): Do best-effort to initialize composite project when HTML is opened first
An upstream TypeScript performance-motivated change (
https://github.com/microsoft/TypeScript/commit/02b07a1788c8ec5b47e51b11955e72ee6d175f4b) made it so composite projects are not opened unless the opened file is in the includes/files list of the `tsconfig`. This means that external HTML files do not trigger project loading when they are the first and only file opened.

As a best-effort, we attempt to open a TS file with the same name, which
will most often be the file for the component of the external template.
This is not guaranteed to be the case, so the only way to guarantee this
works correctly would be for developers to add the `**/*.html` files to
the `includes` list of their `tsconfig[.lib].json`.

resolves #2149